feat: use Z3 FP for float and double

+++ src/main/java/nl/uu/maze/example/ExampleClass.java
@@ -5,7 +5,7 @@ package nl.uu.maze.example;
  */
 public class ExampleClass {
 
-    public static void print(long x) {
+    public static void print(float x) {
         if (x > 5) {
             System.out.println("Positive");
         } else {
